requirements.txt
#requirements.txt

imageio==2.2.0
ipywidgets==7.1.1
matplotlib==3.1.1
mpld3==0.3
numpy==1.14.5
pandas==0.25.1
Pillow==6.1.0
pydicom==1.3.0
requests==2.22.0
scikit-image==0.13.1
scikit-learn==0.19.1
scipy==1.3.1
seaborn==0.9.0
SimpleITK==1.2.2
six==1.12.0
sklearn==0.0